Streaming error fixed && Livestream not showing yet

+++ Backend/upload_images.py
@@ -14,25 +14,28 @@ def upload_images(db,bucket):
     """
     m=0
     while True:
-        images = os.listdir(os.path.join(path_images,'images'))
-        for path in images:
-            loc = path.split('.')[0]
-            loc = loc + '.txt'
-            image_path = os.path.join(path_images,'images',path)
-            # Local path to the image you want to upload
-            local_location_path = os.path.join(path_images,'location',loc)
-            with open(local_location_path,'r') as f:
-                location = f.readline()
-                longi, lati = location.split(" ")
-            # Path in Firebase Storage where you want to store the image
-            # upload location with image name as id
-            firebase_storage_path = f"images/hole{m}.jpg"
-            data = {"longi": longi, "lati": lati}
-            db.collection("streamlocation").document(f"hole{m}.jpg").set(data)
-            # Upload the image to Firebase Storage
-            blob = bucket.blob(firebase_storage_path)
-            blob.upload_from_filename(image_path)
-            print(f"Image uploaded to Firebase Storage: {firebase_storage_path}")
-            os.remove(image_path)
-            os.remove(local_location_path)
-            m+=1
+        images = None
+        if(os.path.exists(os.path.join(path_images,'images'))):
+            images = os.listdir(os.path.join(path_images,'images'))
+        if images!=None:
+            for path in images:
+                loc = path.split('.')[0]
+                loc = loc + '.txt'
+                image_path = os.path.join(path_images,'images',path)
+                # Local path to the image you want to upload
+                local_location_path = os.path.join(path_images,'location',loc)
+                with open(local_location_path,'r') as f:
+                    location = f.readline()
+                    longi, lati = location.split(" ")
+                # Path in Firebase Storage where you want to store the image
+                # upload location with image name as id
+                firebase_storage_path = f"images/hole{m}.jpg"
+                data = {"longi": longi, "lati": lati}
+                db.collection("streamlocation").document(f"hole{m}.jpg").set(data)
+                # Upload the image to Firebase Storage
+                blob = bucket.blob(firebase_storage_path)
+                blob.upload_from_filename(image_path)
+                print(f"Image uploaded to Firebase Storage: {firebase_storage_path}")
+                os.remove(image_path)
+                os.remove(local_location_path)
+                m+=1
